Move related item query to its consumer
