+++ client/src/graphql/generated.ts
@@ -16674,104 +16674,6 @@ export type GQLJobFieldsFragment = {
       };
 };
 
-export type GQLGetRelatedItemsQueryVariables = Exact<{
-  itemIdentifiers:
-    ReadonlyArray<GQLItemIdentifierInput> | GQLItemIdentifierInput;
-}>;
-
-export type GQLGetRelatedItemsQuery = {
-  readonly __typename: 'Query';
-  readonly latestItemSubmissions: ReadonlyArray<
-    | {
-        readonly __typename: 'ContentItem';
-        readonly id: string;
-        readonly submissionId: string;
-        readonly submissionTime?: Date | string | null;
-        readonly data: JsonObject;
-        readonly type: {
-          readonly __typename: 'ContentItemType';
-          readonly id: string;
-          readonly name: string;
-          readonly baseFields: ReadonlyArray<{
-            readonly __typename: 'BaseField';
-            readonly name: string;
-            readonly type: GQLFieldType;
-            readonly required: boolean;
-            readonly container?: {
-              readonly __typename: 'Container';
-              readonly containerType: GQLContainerType;
-              readonly keyScalarType?: GQLScalarType | null;
-              readonly valueScalarType: GQLScalarType;
-            } | null;
-          }>;
-          readonly schemaFieldRoles: {
-            readonly __typename: 'ContentSchemaFieldRoles';
-            readonly displayName?: string | null;
-          };
-        };
-      }
-    | {
-        readonly __typename: 'ThreadItem';
-        readonly id: string;
-        readonly submissionId: string;
-        readonly submissionTime?: Date | string | null;
-        readonly data: JsonObject;
-        readonly type: {
-          readonly __typename: 'ThreadItemType';
-          readonly id: string;
-          readonly name: string;
-          readonly baseFields: ReadonlyArray<{
-            readonly __typename: 'BaseField';
-            readonly name: string;
-            readonly type: GQLFieldType;
-            readonly required: boolean;
-            readonly container?: {
-              readonly __typename: 'Container';
-              readonly containerType: GQLContainerType;
-              readonly keyScalarType?: GQLScalarType | null;
-              readonly valueScalarType: GQLScalarType;
-            } | null;
-          }>;
-          readonly schemaFieldRoles: {
-            readonly __typename: 'ThreadSchemaFieldRoles';
-            readonly displayName?: string | null;
-          };
-        };
-      }
-    | {
-        readonly __typename: 'UserItem';
-        readonly id: string;
-        readonly submissionId: string;
-        readonly submissionTime?: Date | string | null;
-        readonly data: JsonObject;
-        readonly type: {
-          readonly __typename: 'UserItemType';
-          readonly id: string;
-          readonly name: string;
-          readonly baseFields: ReadonlyArray<{
-            readonly __typename: 'BaseField';
-            readonly name: string;
-            readonly type: GQLFieldType;
-            readonly required: boolean;
-            readonly container?: {
-              readonly __typename: 'Container';
-              readonly containerType: GQLContainerType;
-              readonly keyScalarType?: GQLScalarType | null;
-              readonly valueScalarType: GQLScalarType;
-            } | null;
-          }>;
-          readonly schemaFieldRoles: {
-            readonly __typename: 'UserSchemaFieldRoles';
-            readonly displayName?: string | null;
-            readonly createdAt?: string | null;
-            readonly profileIcon?: string | null;
-            readonly backgroundImage?: string | null;
-          };
-        };
-      }
-  >;
-};
-
 export type GQLManualReviewJobCommentFieldsFragment = {
   readonly __typename: 'ManualReviewJobComment';
   readonly id: string;
@@ -16883,6 +16785,104 @@ export type GQLGetThreadHistoryQuery = {
       | { readonly __typename: 'ThreadItem' }
       | { readonly __typename: 'UserItem' };
   }>;
+};
+
+export type GQLGetRelatedItemsQueryVariables = Exact<{
+  itemIdentifiers:
+    ReadonlyArray<GQLItemIdentifierInput> | GQLItemIdentifierInput;
+}>;
+
+export type GQLGetRelatedItemsQuery = {
+  readonly __typename: 'Query';
+  readonly latestItemSubmissions: ReadonlyArray<
+    | {
+        readonly __typename: 'ContentItem';
+        readonly id: string;
+        readonly submissionId: string;
+        readonly submissionTime?: Date | string | null;
+        readonly data: JsonObject;
+        readonly type: {
+          readonly __typename: 'ContentItemType';
+          readonly id: string;
+          readonly name: string;
+          readonly baseFields: ReadonlyArray<{
+            readonly __typename: 'BaseField';
+            readonly name: string;
+            readonly type: GQLFieldType;
+            readonly required: boolean;
+            readonly container?: {
+              readonly __typename: 'Container';
+              readonly containerType: GQLContainerType;
+              readonly keyScalarType?: GQLScalarType | null;
+              readonly valueScalarType: GQLScalarType;
+            } | null;
+          }>;
+          readonly schemaFieldRoles: {
+            readonly __typename: 'ContentSchemaFieldRoles';
+            readonly displayName?: string | null;
+          };
+        };
+      }
+    | {
+        readonly __typename: 'ThreadItem';
+        readonly id: string;
+        readonly submissionId: string;
+        readonly submissionTime?: Date | string | null;
+        readonly data: JsonObject;
+        readonly type: {
+          readonly __typename: 'ThreadItemType';
+          readonly id: string;
+          readonly name: string;
+          readonly baseFields: ReadonlyArray<{
+            readonly __typename: 'BaseField';
+            readonly name: string;
+            readonly type: GQLFieldType;
+            readonly required: boolean;
+            readonly container?: {
+              readonly __typename: 'Container';
+              readonly containerType: GQLContainerType;
+              readonly keyScalarType?: GQLScalarType | null;
+              readonly valueScalarType: GQLScalarType;
+            } | null;
+          }>;
+          readonly schemaFieldRoles: {
+            readonly __typename: 'ThreadSchemaFieldRoles';
+            readonly displayName?: string | null;
+          };
+        };
+      }
+    | {
+        readonly __typename: 'UserItem';
+        readonly id: string;
+        readonly submissionId: string;
+        readonly submissionTime?: Date | string | null;
+        readonly data: JsonObject;
+        readonly type: {
+          readonly __typename: 'UserItemType';
+          readonly id: string;
+          readonly name: string;
+          readonly baseFields: ReadonlyArray<{
+            readonly __typename: 'BaseField';
+            readonly name: string;
+            readonly type: GQLFieldType;
+            readonly required: boolean;
+            readonly container?: {
+              readonly __typename: 'Container';
+              readonly containerType: GQLContainerType;
+              readonly keyScalarType?: GQLScalarType | null;
+              readonly valueScalarType: GQLScalarType;
+            } | null;
+          }>;
+          readonly schemaFieldRoles: {
+            readonly __typename: 'UserSchemaFieldRoles';
+            readonly displayName?: string | null;
+            readonly createdAt?: string | null;
+            readonly profileIcon?: string | null;
+            readonly backgroundImage?: string | null;
+          };
+        };
+      }
+  >;
 };
 
 export type GQLItemTypeHiddenFieldsQueryVariables = Exact<{
@@ -34964,180 +34964,6 @@ export type GQLReleaseJobLockMutationOptions = Apollo.BaseMutationOptions<
   GQLReleaseJobLockMutation,
   GQLReleaseJobLockMutationVariables
 >;
-export const GQLGetRelatedItemsDocument = gql`
-  query getRelatedItems($itemIdentifiers: [ItemIdentifierInput!]!) {
-    latestItemSubmissions(itemIdentifiers: $itemIdentifiers) {
-      ... on UserItem {
-        id
-        submissionId
-        submissionTime
-        data
-        type {
-          id
-          name
-          baseFields {
-            name
-            type
-            required
-            container {
-              containerType
-              keyScalarType
-              valueScalarType
-            }
-          }
-          schemaFieldRoles {
-            displayName
-            createdAt
-            profileIcon
-            backgroundImage
-          }
-        }
-      }
-      ... on ContentItem {
-        id
-        submissionId
-        submissionTime
-        data
-        type {
-          id
-          name
-          baseFields {
-            name
-            type
-            required
-            container {
-              containerType
-              keyScalarType
-              valueScalarType
-            }
-          }
-          schemaFieldRoles {
-            displayName
-          }
-        }
-      }
-      ... on ThreadItem {
-        id
-        submissionId
-        submissionTime
-        data
-        type {
-          id
-          name
-          baseFields {
-            name
-            type
-            required
-            container {
-              containerType
-              keyScalarType
-              valueScalarType
-            }
-          }
-          schemaFieldRoles {
-            displayName
-          }
-        }
-      }
-    }
-  }
-`;
-
-/**
- * __useGQLGetRelatedItemsQuery__
- *
- * To run a query within a React component, call `useGQLGetRelatedItemsQuery` and pass it any options that fit your needs.
- * When your component renders, `useGQLGetRelatedItemsQuery` returns an object from Apollo Client that contains loading, error, and data properties
- * you can use to render your UI.
- *
- * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
- *
- * @example
- * const { data, loading, error } = useGQLGetRelatedItemsQuery({
- *   variables: {
- *      itemIdentifiers: // value for 'itemIdentifiers'
- *   },
- * });
- */
-export function useGQLGetRelatedItemsQuery(
-  baseOptions: Apollo.QueryHookOptions<
-    GQLGetRelatedItemsQuery,
-    GQLGetRelatedItemsQueryVariables
-  > &
-    (
-      | { variables: GQLGetRelatedItemsQueryVariables; skip?: boolean }
-      | { skip: boolean }
-    ),
-) {
-  const options = { ...defaultOptions, ...baseOptions };
-  return Apollo.useQuery<
-    GQLGetRelatedItemsQuery,
-    GQLGetRelatedItemsQueryVariables
-  >(GQLGetRelatedItemsDocument, options);
-}
-export function useGQLGetRelatedItemsLazyQuery(
-  baseOptions?: Apollo.LazyQueryHookOptions<
-    GQLGetRelatedItemsQuery,
-    GQLGetRelatedItemsQueryVariables
-  >,
-) {
-  const options = { ...defaultOptions, ...baseOptions };
-  return Apollo.useLazyQuery<
-    GQLGetRelatedItemsQuery,
-    GQLGetRelatedItemsQueryVariables
-  >(GQLGetRelatedItemsDocument, options);
-}
-// @ts-ignore
-export function useGQLGetRelatedItemsSuspenseQuery(
-  baseOptions?: Apollo.SuspenseQueryHookOptions<
-    GQLGetRelatedItemsQuery,
-    GQLGetRelatedItemsQueryVariables
-  >,
-): Apollo.UseSuspenseQueryResult<
-  GQLGetRelatedItemsQuery,
-  GQLGetRelatedItemsQueryVariables
->;
-export function useGQLGetRelatedItemsSuspenseQuery(
-  baseOptions?:
-    | Apollo.SkipToken
-    | Apollo.SuspenseQueryHookOptions<
-        GQLGetRelatedItemsQuery,
-        GQLGetRelatedItemsQueryVariables
-      >,
-): Apollo.UseSuspenseQueryResult<
-  GQLGetRelatedItemsQuery | undefined,
-  GQLGetRelatedItemsQueryVariables
->;
-export function useGQLGetRelatedItemsSuspenseQuery(
-  baseOptions?:
-    | Apollo.SkipToken
-    | Apollo.SuspenseQueryHookOptions<
-        GQLGetRelatedItemsQuery,
-        GQLGetRelatedItemsQueryVariables
-      >,
-) {
-  const options =
-    baseOptions === Apollo.skipToken
-      ? baseOptions
-      : { ...defaultOptions, ...baseOptions };
-  return Apollo.useSuspenseQuery<
-    GQLGetRelatedItemsQuery,
-    GQLGetRelatedItemsQueryVariables
-  >(GQLGetRelatedItemsDocument, options);
-}
-export type GQLGetRelatedItemsQueryHookResult = ReturnType<
-  typeof useGQLGetRelatedItemsQuery
->;
-export type GQLGetRelatedItemsLazyQueryHookResult = ReturnType<
-  typeof useGQLGetRelatedItemsLazyQuery
->;
-export type GQLGetRelatedItemsSuspenseQueryHookResult = ReturnType<
-  typeof useGQLGetRelatedItemsSuspenseQuery
->;
-export type GQLGetRelatedItemsQueryResult = Apollo.QueryResult<
-  GQLGetRelatedItemsQuery,
-  GQLGetRelatedItemsQueryVariables
->;
 export const GQLGetCommentsForJobDocument = gql`
   query GetCommentsForJob($jobId: ID!) {
     getCommentsForJob(jobId: $jobId) {
@@ -35485,6 +35311,180 @@ export type GQLGetThreadHistorySuspenseQueryHookResult = ReturnType<
 export type GQLGetThreadHistoryQueryResult = Apollo.QueryResult<
   GQLGetThreadHistoryQuery,
   GQLGetThreadHistoryQueryVariables
+>;
+export const GQLGetRelatedItemsDocument = gql`
+  query getRelatedItems($itemIdentifiers: [ItemIdentifierInput!]!) {
+    latestItemSubmissions(itemIdentifiers: $itemIdentifiers) {
+      ... on UserItem {
+        id
+        submissionId
+        submissionTime
+        data
+        type {
+          id
+          name
+          baseFields {
+            name
+            type
+            required
+            container {
+              containerType
+              keyScalarType
+              valueScalarType
+            }
+          }
+          schemaFieldRoles {
+            displayName
+            createdAt
+            profileIcon
+            backgroundImage
+          }
+        }
+      }
+      ... on ContentItem {
+        id
+        submissionId
+        submissionTime
+        data
+        type {
+          id
+          name
+          baseFields {
+            name
+            type
+            required
+            container {
+              containerType
+              keyScalarType
+              valueScalarType
+            }
+          }
+          schemaFieldRoles {
+            displayName
+          }
+        }
+      }
+      ... on ThreadItem {
+        id
+        submissionId
+        submissionTime
+        data
+        type {
+          id
+          name
+          baseFields {
+            name
+            type
+            required
+            container {
+              containerType
+              keyScalarType
+              valueScalarType
+            }
+          }
+          schemaFieldRoles {
+            displayName
+          }
+        }
+      }
+    }
+  }
+`;
+
+/**
+ * __useGQLGetRelatedItemsQuery__
+ *
+ * To run a query within a React component, call `useGQLGetRelatedItemsQuery` and pass it any options that fit your needs.
+ * When your component renders, `useGQLGetRelatedItemsQuery` returns an object from Apollo Client that contains loading, error, and data properties
+ * you can use to render your UI.
+ *
+ * @param baseOptions options that will be passed into the query, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options;
+ *
+ * @example
+ * const { data, loading, error } = useGQLGetRelatedItemsQuery({
+ *   variables: {
+ *      itemIdentifiers: // value for 'itemIdentifiers'
+ *   },
+ * });
+ */
+export function useGQLGetRelatedItemsQuery(
+  baseOptions: Apollo.QueryHookOptions<
+    GQLGetRelatedItemsQuery,
+    GQLGetRelatedItemsQueryVariables
+  > &
+    (
+      | { variables: GQLGetRelatedItemsQueryVariables; skip?: boolean }
+      | { skip: boolean }
+    ),
+) {
+  const options = { ...defaultOptions, ...baseOptions };
+  return Apollo.useQuery<
+    GQLGetRelatedItemsQuery,
+    GQLGetRelatedItemsQueryVariables
+  >(GQLGetRelatedItemsDocument, options);
+}
+export function useGQLGetRelatedItemsLazyQuery(
+  baseOptions?: Apollo.LazyQueryHookOptions<
+    GQLGetRelatedItemsQuery,
+    GQLGetRelatedItemsQueryVariables
+  >,
+) {
+  const options = { ...defaultOptions, ...baseOptions };
+  return Apollo.useLazyQuery<
+    GQLGetRelatedItemsQuery,
+    GQLGetRelatedItemsQueryVariables
+  >(GQLGetRelatedItemsDocument, options);
+}
+// @ts-ignore
+export function useGQLGetRelatedItemsSuspenseQuery(
+  baseOptions?: Apollo.SuspenseQueryHookOptions<
+    GQLGetRelatedItemsQuery,
+    GQLGetRelatedItemsQueryVariables
+  >,
+): Apollo.UseSuspenseQueryResult<
+  GQLGetRelatedItemsQuery,
+  GQLGetRelatedItemsQueryVariables
+>;
+export function useGQLGetRelatedItemsSuspenseQuery(
+  baseOptions?:
+    | Apollo.SkipToken
+    | Apollo.SuspenseQueryHookOptions<
+        GQLGetRelatedItemsQuery,
+        GQLGetRelatedItemsQueryVariables
+      >,
+): Apollo.UseSuspenseQueryResult<
+  GQLGetRelatedItemsQuery | undefined,
+  GQLGetRelatedItemsQueryVariables
+>;
+export function useGQLGetRelatedItemsSuspenseQuery(
+  baseOptions?:
+    | Apollo.SkipToken
+    | Apollo.SuspenseQueryHookOptions<
+        GQLGetRelatedItemsQuery,
+        GQLGetRelatedItemsQueryVariables
+      >,
+) {
+  const options =
+    baseOptions === Apollo.skipToken
+      ? baseOptions
+      : { ...defaultOptions, ...baseOptions };
+  return Apollo.useSuspenseQuery<
+    GQLGetRelatedItemsQuery,
+    GQLGetRelatedItemsQueryVariables
+  >(GQLGetRelatedItemsDocument, options);
+}
+export type GQLGetRelatedItemsQueryHookResult = ReturnType<
+  typeof useGQLGetRelatedItemsQuery
+>;
+export type GQLGetRelatedItemsLazyQueryHookResult = ReturnType<
+  typeof useGQLGetRelatedItemsLazyQuery
+>;
+export type GQLGetRelatedItemsSuspenseQueryHookResult = ReturnType<
+  typeof useGQLGetRelatedItemsSuspenseQuery
+>;
+export type GQLGetRelatedItemsQueryResult = Apollo.QueryResult<
+  GQLGetRelatedItemsQuery,
+  GQLGetRelatedItemsQueryVariables
 >;
 export const GQLItemTypeHiddenFieldsDocument = gql`
   query ItemTypeHiddenFields {
@@ -45383,9 +45383,9 @@ export const namedOperations = {
     GetDecidedJob: 'GetDecidedJob',
     ManualReviewSafetySettings: 'ManualReviewSafetySettings',
     ManualReviewJobInfo: 'ManualReviewJobInfo',
-    getRelatedItems: 'getRelatedItems',
     GetCommentsForJob: 'GetCommentsForJob',
     getThreadHistory: 'getThreadHistory',
+    getRelatedItems: 'getRelatedItems',
     ItemTypeHiddenFields: 'ItemTypeHiddenFields',
     AllManualReviewQueues: 'AllManualReviewQueues',
     getLatestUserSubmittedItemsWithThreads:
