+++ client/src/webpages/dashboard/mrt/manual_review_job/v2/ContentRelatedItemComponent.tsx
@@ -1,89 +1,7 @@
 import { GQLContentItem } from '@/graphql/generated';
 import { ItemTypeFieldFieldData } from '@/webpages/dashboard/item_types/itemTypeUtils';
-import { gql } from '@apollo/client';
 
 import FieldsComponent from './ManualReviewJobFieldsComponent';
-
-// ManualReviewJobFieldsComponent uses this operation to resolve RELATED_ITEM
-// fields.
-gql`
-  query getRelatedItems($itemIdentifiers: [ItemIdentifierInput!]!) {
-    latestItemSubmissions(itemIdentifiers: $itemIdentifiers) {
-      ... on UserItem {
-        id
-        submissionId
-        submissionTime
-        data
-        type {
-          id
-          name
-          baseFields {
-            name
-            type
-            required
-            container {
-              containerType
-              keyScalarType
-              valueScalarType
-            }
-          }
-          schemaFieldRoles {
-            displayName
-            createdAt
-            profileIcon
-            backgroundImage
-          }
-        }
-      }
-      ... on ContentItem {
-        id
-        submissionId
-        submissionTime
-        data
-        type {
-          id
-          name
-          baseFields {
-            name
-            type
-            required
-            container {
-              containerType
-              keyScalarType
-              valueScalarType
-            }
-          }
-          schemaFieldRoles {
-            displayName
-          }
-        }
-      }
-      ... on ThreadItem {
-        id
-        submissionId
-        submissionTime
-        data
-        type {
-          id
-          name
-          baseFields {
-            name
-            type
-            required
-            container {
-              containerType
-              keyScalarType
-              valueScalarType
-            }
-          }
-          schemaFieldRoles {
-            displayName
-          }
-        }
-      }
-    }
-  }
-`;
 
 type LoadedContentItem = Pick<GQLContentItem, 'data'> & {
   type: Pick<GQLContentItem['type'], 'id' | 'baseFields'>;
